Heavily refactored LocalGameController
