+++ src/main/java/ui/controller/ScorePaneController.java
@@ -21,9 +21,9 @@ import event.EventBus;
 import event.ScoreEvent;
 import javafx.beans.value.ObservableValue;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Node;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.Separator;
@@ -32,11 +32,10 @@ import javafx.scene.layout.VBox;
 import logic.score.Scorer;
 import util.StoneColour;
 
-import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class ScorePaneController implements Initializable {
+public class ScorePaneController extends BuildableController implements Initializable {
 
 	@FXML
 	private Pane scorePane;
@@ -51,17 +50,13 @@ public class ScorePaneController implements Initializable {
 
 	private Scorer scorer;
 
-	static FXMLLoader getScorePaneLoader() {
-		FXMLLoader loader = new FXMLLoader(ScorePaneController.class.getResource("/scorePane.fxml"));
+	ScorePaneController() {
+		scorer = null;
+	}
 
-		try {
-			loader.load();
-		}
-		catch (IOException e) {
-			e.printStackTrace();
-		}
-
-		return loader;
+	@Override
+	protected String getResourcePath() {
+		return "/scorePane.fxml";
 	}
 
 	public void setScorer(Scorer scorer) {
@@ -71,6 +66,8 @@ public class ScorePaneController implements Initializable {
 	@Override
 	public void initialize(URL url, ResourceBundle resourceBundle) {
 		scorePane.widthProperty().addListener(this::resizeScore);
+		enableButtons();
+		setButtonActions();
 		setVisible(false);
 
 		EventBus.addEventHandler(ScoreEvent.ANY, this::updateScore);
@@ -85,18 +82,50 @@ public class ScorePaneController implements Initializable {
 		whiteDone.setDisable(false);
 	}
 
-	void setDoneScoring(Runnable callback) {
+	private void setButtonActions() {
 		blackDone.setOnAction(event -> {
 			blackDone.setDisable(true);
 			if ( whiteDone.isDisabled() )
-				callback.run();
+				fireDoneScoring();
 		});
 
 		whiteDone.setOnAction(event -> {
 			whiteDone.setDisable(true);
 			if ( blackDone.isDisabled() )
-				callback.run();
+				fireDoneScoring();
 		});
+	}
+
+	private void fireDoneScoring() {
+		EventBus bus = EventBus.getInstance();
+		ScoreEvent doneScoring = new ScoreEvent(scorer, bus, ScoreEvent.DONE);
+		bus.fireEvent(doneScoring);
+
+		displayFinalScore(scorer.getScore());
+	}
+
+	private void displayFinalScore(double finalScore) {
+		Alert alert = new Alert(Alert.AlertType.INFORMATION);
+		alert.setTitle("Game Over");
+		if ( finalScore == 0 ) {
+			alert.setContentText("Game ends in a draw.");
+			alert.showAndWait();
+			return;
+		}
+
+		String message;
+
+		if ( finalScore > 0 )
+			message = "Black";
+		else
+			message = "White";
+
+		finalScore = Math.abs(finalScore);
+
+		message += " wins by " + Double.toString(finalScore) + " points.";
+
+		alert.setContentText(message);
+		alert.showAndWait();
 	}
 
 	private void updateScore(ScoreEvent event) {
