+++ src/main/resources/gameMenu.fxml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.*?>
+<VBox xmlns:fx="http://javafx.com/fxml"
+      xmlns="http://javafx.com/javafx"
+      fx:controller="ui.controller.GameMenuController">
+    <AnchorPane>
+        <Button fx:id="passButton" mnemonicParsing="false" text="Pass" AnchorPane.leftAnchor="0"
+                AnchorPane.rightAnchor="0"/>
+    </AnchorPane>
+    <AnchorPane>
+        <Button fx:id="undoButton" mnemonicParsing="false" text="Undo" AnchorPane.leftAnchor="0"
+                AnchorPane.rightAnchor="0"/>
+    </AnchorPane>
+    <AnchorPane>
+
+        <Button fx:id="saveButton" mnemonicParsing="false" text="Save as SGF" AnchorPane.leftAnchor="0"
+                AnchorPane.rightAnchor="0"/>
+    </AnchorPane>
+</VBox>
