+++ src/main/resources/localGame.fxml
@@ -1,28 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.SplitPane?>
-<?import javafx.scene.layout.*?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
 <SplitPane xmlns:fx="http://javafx.com/fxml/1" dividerPositions="0.7" prefHeight="1000000" prefWidth="1000000"
            xmlns="http://javafx.com/javafx/8.0.111" fx:controller="ui.controller.BoardController">
 
     <Pane fx:id="boardPane" minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0"/>
-    <VBox fx:id="sideBar" prefHeight="200.0" prefWidth="100.0">
-
-        <AnchorPane>
-            <Button fx:id="passButton" mnemonicParsing="false" text="Pass" AnchorPane.leftAnchor="0"
-                    AnchorPane.rightAnchor="0"/>
-        </AnchorPane>
-        <AnchorPane>
-            <Button fx:id="undoButton" mnemonicParsing="false" text="Undo" AnchorPane.leftAnchor="0"
-                    AnchorPane.rightAnchor="0"/>
-        </AnchorPane>
-        <AnchorPane>
-            <Button fx:id="saveButton" mnemonicParsing="false" text="Save as SGF" AnchorPane.leftAnchor="0"
-                    AnchorPane.rightAnchor="0"/>
-        </AnchorPane>
-
-    </VBox>
+    <VBox fx:id="sideBar" prefHeight="200.0" prefWidth="100.0"/>
 
 </SplitPane>
